bible-outline-enhanced-backend/requirements.txt
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-docx==1.2.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
# Additional dependencies for deployment
requests==2.31.0
beautifulsoup4==4.12.2
APScheduler==3.10.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
PyMuPDF==1.23.5
